requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
werkzeug==3.0.1
pandas==2.1.3
numpy==1.26.2
torch==2.1.1
scikit-learn==1.3.2
jupyter==1.0.0
notebook==7.0.6
